util page component removed
