+++ src/views/Util.vue
@@ -1,17 +1,45 @@
 <template>
-  <div class="util">
-    <UtilPage />
-  </div>
+  <v-container class="mt-5">
+    <div class="util-info-box">
+      <h5>Your User Agent</h5>
+      <h4 id="user-agent-value"></h4>
+    </div>
+    <div class="util-info-box">
+      <h5>Your IP Info</h5>
+      <h4 id="ip-addr-value">👀</h4>
+    </div>
+  </v-container>
 </template>
 
 <script>
-// @ is an alias to /src
-import UtilPage from '@/components/UtilPage.vue'
+import axios from 'axios'
 
 export default {
   name: 'Util',
-  components: {
-    UtilPage
+  mounted: function () {
+    document.getElementById('user-agent-value').innerHTML = navigator.userAgent
+
+    axios
+      .get(process.env.VUE_APP_API_ENDPOINT + '/myip')
+      .then(response => {
+        if (response === null || response.data === null) {
+          console.error('received null response / data messages')
+          return
+        }
+        document.getElementById('ip-addr-value').innerHTML = response.data
+      })
+      .catch(error => {
+        console.log(error)
+      })
   }
 }
 </script>
+
+<style scoped>
+.util-info-box {
+    background-color: #26A69A;
+    margin-left: 10%;
+    margin-right: 10%;
+    border-radius: 5px;
+}
+</style>
